still working on calendar events, started styling and functionality after looking through docs
