+++ client/src/components/AuthViews/DisplayPage/RightSideCol.jsx
@@ -7,7 +7,7 @@ import ChatIcon from "@mui/icons-material/Chat";
 import { accountContext } from "../../Contexts/appContext";
 import { IndividualChats } from "../ChatViews/IndividualChat";
 import { useNavigate } from "react-router-dom";
-import Test from "./Test";
+import { EventCalendar } from "./EventCalendar";
 
 
 export const RightSideCol = () => {
@@ -76,7 +76,7 @@ export const RightSideCol = () => {
 
   return (
     <div className="right_column_wrapper">
-      <Test/>
+      <EventCalendar/>
       <div className="popular_container">
         <h2
           style={{
